test: move partial batch test into core tests
